src/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "av<14",  # Requiere FFmpeg 7, pero Debian 12 tiene FFmpeg 5.x - usar version compatible
    "azure-ai-agents>=1.0.0b10",
    "azure-ai-evaluation>=1.5.0",
    "azure-ai-inference==1.0.0b9",
    "azure-ai-projects==1.0.0b11",
    "azure-cosmos>=4.9.0",
    "azure-identity>=1.21.0",
    "azure-keyvault-secrets>=4.4.0",
    "azure-monitor-events-extension>=0.1.0",
    "azure-monitor-opentelemetry>=1.6.8",
    "azure-search-documents>=11.5.2",
    "boto3>=1.34.0",
    "fastapi>=0.115.12",
    "google-cloud-firestore>=2.19.0",
    "httpx>=0.28.1",
    "numpy>=2.2.5",
    "openai==1.84.0",
    "opentelemetry-api>=1.31.1",
    "opentelemetry-exporter-otlp-proto-grpc>=1.31.1",
    "opentelemetry-exporter-otlp-proto-http>=1.31.1",
    "opentelemetry-instrumentation-fastapi>=0.52b1",
    "opentelemetry-instrumentation-openai>=0.39.2",
    "opentelemetry-sdk>=1.31.1",
    "pytest>=8.2,<9",
    "pytest-asyncio==0.24.0",
    "pytest-cov==5.0.0",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "semantic-kernel[azure]==1.32.2",
    "urllib3>=1.26.0,<2",
    "chardet>=3.0.2,<6",
    "uvicorn>=0.34.2",
]

[dependency-groups]
dev = [
    "ruff>=0.15.5",
]
